cleanup help page assets & content
