+++ app/style/Header.less
@@ -65,9 +65,6 @@
   &.accounts {
     background-image: var(--header-accounts);
   }
-  &.help {
-    background-image: var(--header-help);
-  }
   &.security {
     background-image: var(--header-security);
   }
