+++ app/style/themes/darkTheme.js
@@ -138,7 +138,6 @@ const darkTheme = {
   "menu-overview": "url('style/icons/overviewActiveDark.svg')",
   "menu-privacy": "url('style/icons/privacyDark.svg')",
   "menu-tickets": "url('style/icons/ticketsActiveDark.svg')",
-  "menu-help": "url('style/icons/helpActiveDark.svg')",
   "menu-ln": "url('style/icons/lightningDark.svg')",
   "menu-trezor": "url('style/icons/trezorActiveDark.svg')",
   "menu-hamburger": "url('style/icons/hamburgerDark.svg')",
@@ -165,7 +164,6 @@ const darkTheme = {
   "stakey-privacy-standard": "url('style/icons/privacyStandardDark.svg')",
   "stakey-privacy-disabled": "url('style/icons/privacyDisabledDark.svg')",
   "stakey-privacy-custom": "url('style/icons/privacyCustomDark.svg')",
-  "header-help": "url('style/icons/helpHeaderDark.svg')",
   "header-transactions": "url('style/icons/transactionsHeaderDark.svg')",
   "header-governance": "url('style/icons/governanceHeaderDark.svg')",
   "header-tickets": "url('style/icons/ticketsHeaderDark.svg')",
