+++ app/style/themes/lightTheme.js
@@ -136,7 +136,6 @@ const lightTheme = {
   "menu-overview": "url('style/icons/overview-active.png')",
   "menu-privacy": "url('style/icons/privacy.svg')",
   "menu-tickets": "url('style/icons/tickets-active.png')",
-  "menu-help": "url('style/icons/help-active.png')",
   "menu-ln": "url('style/icons/lightning.svg')",
   "menu-trezor": "url('style/icons/trezorActive.svg')",
   "menu-hamburger": "url('style/icons/hamburger.svg')",
@@ -163,7 +162,6 @@ const lightTheme = {
   "stakey-privacy-standard": "url('style/icons/privacyStandard.svg')",
   "stakey-privacy-disabled": "url('style/icons/privacyDisabled.svg')",
   "stakey-privacy-custom": "url('style/icons/privacyCustom.svg')",
-  "header-help": "url('style/icons/helpHeader.svg')",
   "header-transactions": "url('style/icons/transactionsHeader.svg')",
   "header-governance": "url('style/icons/governanceHeader.svg')",
   "header-tickets": "url('style/icons/ticketsHeader.svg')",
